refactor: 리팩터링 및 버그수정
