--- android/Bottari/presentation/src/main/java/com/bottari/presentation/model/TeamChecklistProductUiModel.kt
@@ -1,14 +1,0 @@
-package com.bottari.presentation.model
-
-import android.os.Parcelable
-import com.bottari.presentation.view.checklist.team.checklist.ChecklistType
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class TeamChecklistProductUiModel(
-    val id: Long,
-    val name: String,
-    val isChecked: Boolean,
-    val type: ChecklistType,
-) : Parcelable,
-    TeamChecklistItem
