--- android/Bottari/presentation/src/main/java/com/bottari/presentation/model/TeamChecklistTypeUiModel.kt
@@ -1,9 +1,0 @@
-package com.bottari.presentation.model
-
-import com.bottari.presentation.view.checklist.team.checklist.ChecklistType
-
-data class TeamChecklistTypeUiModel(
-    val type: ChecklistType,
-    val teamChecklistItems: List<TeamChecklistProductUiModel>,
-    var isExpanded: Boolean = true,
-) : TeamChecklistItem
